python script/cascadya_features/requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
psycopg[binary]>=3.1,<4.0
waitress>=3.0,<4.0
